separate buy page components
