+++ ui/page/buy/index.js
@@ -1,17 +1,6 @@
 import { connect } from 'react-redux';
-import { selectGettingNewAddress, selectReceiveAddress, doGetNewAddress } from 'lbry-redux';
-import { selectUserEmail, selectUser } from 'redux/selectors/user';
-import { doUserSetCountry } from 'redux/actions/user';
 import BuyPage from './view';
 
-const select = state => ({
-  receiveAddress: selectReceiveAddress(state),
-  gettingNewAddress: selectGettingNewAddress(state),
-  email: selectUserEmail(state),
-  user: selectUser(state),
-});
+const select = state => ({});
 
-export default connect(select, {
-  doGetNewAddress,
-  doUserSetCountry,
-})(BuyPage);
+export default connect(select, null)(BuyPage);
